add say_jokes.py
